+++ src/refactoring/components/admin/index.ts
@@ -1,0 +1,2 @@
+export * from "./ProductManagementSection";
+export * from "./CouponManagementSection";
